Script Final/requirements.txt
pandas>=2.2.0
numpy>=1.26.0
mysql-connector-python>=9.0.0
python-docx>=1.1.2